refactor(codex-local): relocate Codex auth-merge scripts into the adapter
Phase 2 of the generic asset-lifecycle-seam refactor. Move the
Codex-specific inbound auth-merge assets out of adapter-utils and into the
codex-local adapter, so adapter-utils retains only generic runtime code:

- git mv codex-auth-merge-extract.sh, codex-auth-merge-decision.cjs, and
  codex-auth-merge-scripts.ts to packages/adapters/codex-local/src/server/
  (scripts byte-identical; scripts.ts imports rebased to adapter-utils
  subpaths for shellQuote + the SandboxManagedRuntimeAssetProvision type).
- execute.ts reads buildCodexAuthInboundProvision from the local module.
- Move the Codex-auth test rows into codex-local's codex-auth-merge.test.ts;
  the generic workspace-restore-merge test stays and no longer references Codex.
- Packaging: codex-local build copies the .sh/.cjs into dist/server; drop the
  now-obsolete copy step from the adapter-utils build.

Pure relocation — no change to script contents or to which bytes cross the
sandbox boundary. Inbound behavior unchanged.
