+++ packages/adapters/codex-local/src/server/codex-auth-merge-extract.sh
